move rtc libraries to lib/
